// src/interfaces/IErrorResponse.ts
export default interface IErrorResponse {
  likedAlready?: string;
  cannotUnlike?: string;
  text?: string;
  handle?: string;
  status?: string;
  skills?: string;
  website?: string;
  youtube?: string;
  twitter?: string;
  linkedin?: string;
  facebook?: string;
  instagram?: string;
  title?: string;
  company?: string;
  from?: string;
  school?: string;
  degree?: string;
  fieldOfStudy?: string;
  name?: string;
  email?: string;
  password?: string;
  password2?: string;
  unauthorised?: string;
  postNotFound?: string;
  commentNotFound?: string;
  noProfile?: string;
  noProfiles?: string;
  educationNotFound?: string;
  experienceNotFound?: string;
  emailNotFound?: string;
}